Rename controllers, add postman collection for easy importing
